+++ tests/front/unit/jest/unit.jest.js
@@ -21,8 +21,6 @@ const unitConfig = {
     'src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/pages/create-unit/CreateUnit.tsx',
     'src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/icons',
     'src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/panel/announcement/Image.tsx',
-    'src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/panel/Panel.tsx',
-    'src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/panel/AnnouncementList.tsx',
     'src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/view',
     'src/Akeneo/Pim/Structure/Bundle/Resources/public/js/attribute-option/contexts',
     'src/Akeneo/Pim/Structure/Bundle/Resources/public/js/attribute-option/fetchers',
@@ -50,7 +48,12 @@ const unitConfig = {
   coverageReporters: ['text-summary', 'html'],
   coverageDirectory: '<rootDir>/coverage/',
   coverageThreshold: {
-    global: {
+    '**/Akeneo/Platform/Bundle/UIBundle/**': {
+      statements: 100,
+      functions: 100,
+      lines: 100,
+    },
+    '**/Akeneo/Tool/Bundle/MeasureBundle/**': {
       statements: 100,
       functions: 100,
       lines: 100,